requirements.txt
streamlit==1.26.0
opencv-python==4.8.1.78
numpy==1.24.4
Pillow==10.0.0